+++ back/Journalist.Crm/Ideas/DataModels/IdeaResultSet.cs
@@ -5,7 +5,7 @@ namespace Journalist.Crm.Domain.Ideas.DataModels
 {
     public class IdeaResultSet : ResultSetBase<Idea>
     {
-        public IdeaResultSet(IReadOnlyCollection<Idea> data, int totalItemCount, bool hasNextPage, bool hasPreviousPage) : base(data, totalItemCount, hasNextPage, hasPreviousPage)
+        public IdeaResultSet(IReadOnlyCollection<Idea> data, long totalItemCount, bool hasNextPage, bool hasPreviousPage) : base(data, totalItemCount, hasNextPage, hasPreviousPage)
         {
         }
     }
