+++ back/Journalist.Crm/Pitches/DataModels/PitchResultSet.cs
@@ -5,7 +5,7 @@ namespace Journalist.Crm.Domain.Pitches.DataModels
 {
     public class PitchResultSet : ResultSetBase<Pitch>
     {
-        public PitchResultSet(IReadOnlyCollection<Pitch> data, int totalItemCount, bool hasNextPage, bool hasPreviousPage) : base(data, totalItemCount, hasNextPage, hasPreviousPage)
+        public PitchResultSet(IReadOnlyCollection<Pitch> data, long totalItemCount, bool hasNextPage, bool hasPreviousPage) : base(data, totalItemCount, hasNextPage, hasPreviousPage)
         {
         }
     }
